Configures publication to GitHub Packages
Configures the project to publish artifacts to GitHub Packages
in addition to Maven Central. This allows for easier internal
distribution and consumption of the library. It also fixes the
publication task name for GitHub Packages.

Additionally, the GPG key setup process in the Maven Central
publication workflow is improved by explicitly exporting and using
the signing key, improving reliability and security.