Fixes signing configuration and adds debug info
Fixes the signing configuration to properly use in-memory PGP keys by removing the Maven Central conditional.

Adds a debug step in the publishing workflow to print the length of the Maven Central username and the GPG Key ID to aid in troubleshooting authentication issues.
Also adds stacktrace to the gradle command to help debug.